#10 - Add findById in Student Service Impl

+++ src/main/java/com/hendisantika/service/StudentService.java
@@ -17,11 +17,11 @@ import java.util.Optional;
 public interface StudentService {
     List<Student> getAllStudents();
 
-    Optional<Student> findById(int id);
+    Optional<Student> findById(Long id);
 
     Optional<Student> findByEmail(String email);
 
     Student save(Student std);
 
-    void deleteById(int id);
+    void deleteById(Long id);
 }
